Refactor package cabinet creation in Binder into an ICommand.
In an effort to massively improve the code in the Binder, the gigantic
class is being broken into ICommand based classes. The first is the
package cabinet creation code. This first step of the refactoring simply
extracts the existing code from the Binder into the
CreateCabinetsCommand. Improving the code is the next step.
