+++ src/tools/wix/Bind/CabinetBuilder.cs
