+++ src/tools/wix/Bind/CabinetWorkItem.cs
