+++ src/tools/wix/Binder.cs
@@ -16,7 +16,6 @@ namespace WixToolset
     using System;
     using System.Collections;
     using System.Collections.Generic;
-    using System.Collections.ObjectModel;
     using System.Collections.Specialized;
     using System.ComponentModel;
     using System.Diagnostics;
@@ -26,12 +25,10 @@ namespace WixToolset
     using System.Linq;
     using System.Reflection;
     using System.Runtime.InteropServices;
-    using System.Security.Cryptography;
-    using System.Security.Cryptography.X509Certificates;
     using System.Text;
-    using System.Threading;
     using System.Xml;
     using System.Xml.XPath;
+    using WixToolset.Bind;
     using WixToolset.Cab;
     using WixToolset.CLR.Interop;
     using WixToolset.Data;
@@ -40,7 +37,6 @@ namespace WixToolset
     using WixToolset.MergeMod;
     using WixToolset.Msi;
     using WixToolset.Msi.Interop;
-    using Wix = WixToolset.Data.Serialize;
 
     // TODO: (4.0) Refactor so that these don't need to be copied.
     // Copied verbatim from ext\UtilExtension\wixext\UtilCompiler.cs
@@ -114,8 +110,6 @@ namespace WixToolset
         // Following object handles are needed by the NewCabNamesCallBack callback
         private List<FileTransfer> fileTransfers; // File Transfers for BindDatabase Only and not the one for BindBundle
         private Output output;
-        internal FileSplitCabNamesCallback newCabNamesCallBack;
-        private Dictionary<string, string> lastCabinetAddedToMediaTable; // Key is First Cabinet Name, Value is Last Cabinet Added in the Split Sequence
 
         /// <summary>
         /// Creates an MSI binder.
@@ -139,8 +133,6 @@ namespace WixToolset
 
             // Need fileTransfers handle for NewCabNamesCallBack callback
             this.fileTransfers = new List<FileTransfer>();
-            this.newCabNamesCallBack = NewCabNamesCallBack;
-            this.lastCabinetAddedToMediaTable = new Dictionary<string, string>();
         }
 
         public string ContentsFile { get; set; }
@@ -348,7 +340,6 @@ namespace WixToolset
 
             this.fileManagerCore = new BinderFileManagerCore();
             this.fileManagerCore.CabCachePath = this.CabCachePath;
-            this.fileManagerCore.ReuseCabinets = this.ReuseCabinets;
             this.fileManagerCore.Output = output;
             this.fileManagerCore.TempFilesLocation = this.TempFilesLocation;
             this.fileManagerCore.AddBindPaths(this.BindPaths, BindStage.Normal);
@@ -1743,7 +1734,21 @@ namespace WixToolset
             if (!this.SuppressLayout || OutputType.Module == output.Type)
             {
                 this.core.OnMessage(WixVerboses.CreatingCabinetFiles());
-                uncompressedFileRows = this.CreateCabinetFiles(output, this.fileTransfers, layoutDirectory, compressed, autoMediaAssigner);
+
+                CreateCabinetsCommand command = new CreateCabinetsCommand();
+                command.CabbingThreadCount = this.CabbingThreadCount;
+                command.DefaultCompressionLevel = this.DefaultCompressionLevel;
+                command.Output = output;
+                command.FileManagers = this.fileManagers;
+                command.FileTransfers = this.fileTransfers;
+                command.LayoutDirectory = layoutDirectory;
+                command.Compressed = compressed;
+                command.AutoMediaAssigner = autoMediaAssigner;
+                command.TableDefinitions = this.core.TableDefinitions;
+                command.TempFilesLocation = this.TempFilesLocation;
+                command.Execute();
+
+                uncompressedFileRows = command.UncompressedFileRows;
             }
 
             if (OutputType.Patch == output.Type)
@@ -2364,22 +2369,6 @@ namespace WixToolset
             {
                 throw new InvalidOperationException(); // TODO: something needs to be said here that none of the binder file managers returned a result.
             }
-        }
-
-        private ResolvedCabinet ResolveCabinet(string cabinetPath, IEnumerable<FileRow> fileRows)
-        {
-            ResolvedCabinet resolved = null;
-
-            foreach (IBinderFileManager fileManager in this.fileManagers)
-            {
-                resolved = fileManager.ResolveCabinet(cabinetPath, fileRows);
-                if (null != resolved)
-                {
-                    break;
-                }
-            }
-
-            return resolved;
         }
 
         private string ResolveFile(string source, string type, SourceLineNumber sourceLineNumbers, BindStage bindStage = BindStage.Normal)
@@ -6796,58 +6785,6 @@ namespace WixToolset
         }
 
         /// <summary>
-        /// Creates cabinet files.
-        /// </summary>
-        /// <param name="output">Output to generate image for.</param>
-        /// <param name="fileTransfers">Array of files to be transfered.</param>
-        /// <param name="layoutDirectory">The directory in which the image should be layed out.</param>
-        /// <param name="compressed">Flag if source image should be compressed.</param>
-        /// <returns>The uncompressed file rows.</returns>
-        private RowDictionary<FileRow> CreateCabinetFiles(Output output, List<FileTransfer> fileTransfers, string layoutDirectory, bool compressed, AutoMediaAssigner autoMediaAssigner)
-        {
-            this.SetCabbingThreadCount();
-
-            // Send Binder object to Facilitate NewCabNamesCallBack Callback
-            CabinetBuilder cabinetBuilder = new CabinetBuilder(this.CabbingThreadCount, Marshal.GetFunctionPointerForDelegate(this.newCabNamesCallBack));
-
-            // Supply Compile MediaTemplate Attributes to Cabinet Builder
-            int MaximumCabinetSizeForLargeFileSplitting;
-            int MaximumUncompressedMediaSize;
-            this.GetMediaTemplateAttributes(out MaximumCabinetSizeForLargeFileSplitting, out MaximumUncompressedMediaSize);
-            cabinetBuilder.MaximumCabinetSizeForLargeFileSplitting = MaximumCabinetSizeForLargeFileSplitting;
-            cabinetBuilder.MaximumUncompressedMediaSize = MaximumUncompressedMediaSize;
-
-            foreach (var entry in autoMediaAssigner.Cabinets)
-            {
-                MediaRow mediaRow = entry.Key;
-                List<FileRow> files = entry.Value;
-
-                string cabinetDir = this.ResolveMedia(mediaRow, layoutDirectory);
-
-                CabinetWorkItem cabinetWorkItem = this.CreateCabinetWorkItem(output, cabinetDir, mediaRow, files, fileTransfers);
-                if (null != cabinetWorkItem)
-                {
-                    cabinetBuilder.Enqueue(cabinetWorkItem);
-                }
-            }
-
-            // stop processing if an error previously occurred
-            if (this.core.EncounteredError)
-            {
-                return null;
-            }
-
-            // create queued cabinets with multiple threads
-            cabinetBuilder.CreateQueuedCabinets();
-            if (this.core.EncounteredError)
-            {
-                return null;
-            }
-
-            return autoMediaAssigner.UncompressedFileRows;
-        }
-
-        /// <summary>
         /// Sets the codepage of a database.
         /// </summary>
         /// <param name="db">Database to set codepage into.</param>
@@ -6874,99 +6811,6 @@ namespace WixToolset
                 // the IDT should be valid, so an invalid code page was given
                 throw new WixException(WixErrors.IllegalCodepage(output.Codepage));
             }
-        }
-
-        /// <summary>
-        /// Creates a work item to create a cabinet.
-        /// </summary>
-        /// <param name="output">Output for the current database.</param>
-        /// <param name="cabinetDir">Directory to create cabinet in.</param>
-        /// <param name="mediaRow">MediaRow containing information about the cabinet.</param>
-        /// <param name="fileRows">Collection of files in this cabinet.</param>
-        /// <param name="fileTransfers">Array of files to be transfered.</param>
-        /// <returns>created CabinetWorkItem object</returns>
-        private CabinetWorkItem CreateCabinetWorkItem(Output output, string cabinetDir, MediaRow mediaRow, List<FileRow> fileRows, List<FileTransfer> fileTransfers)
-        {
-            CabinetWorkItem cabinetWorkItem = null;
-            string tempCabinetFileX = Path.Combine(this.TempFilesLocation, mediaRow.Cabinet);
-
-            // check for an empty cabinet
-            if (0 == fileRows.Count)
-            {
-                string cabinetName = mediaRow.Cabinet;
-
-                // remove the leading '#' from the embedded cabinet name to make the warning easier to understand
-                if (cabinetName.StartsWith("#", StringComparison.Ordinal))
-                {
-                    cabinetName = cabinetName.Substring(1);
-                }
-
-                // If building a patch, remind them to run -p for torch.
-                if (OutputType.Patch == output.Type)
-                {
-                    this.core.OnMessage(WixWarnings.EmptyCabinet(mediaRow.SourceLineNumbers, cabinetName, true));
-                }
-                else
-                {
-                    this.core.OnMessage(WixWarnings.EmptyCabinet(mediaRow.SourceLineNumbers, cabinetName));
-                }
-            }
-
-            ResolvedCabinet resolvedCabinet = this.ResolveCabinet(tempCabinetFileX, fileRows);
-
-            // create a cabinet work item if it's not being skipped
-            if (CabinetBuildOption.BuildAndCopy == resolvedCabinet.BuildOption || CabinetBuildOption.BuildAndMove == resolvedCabinet.BuildOption)
-            {
-                int maxThreshold = 0; // default to the threshold for best smartcabbing (makes smallest cabinet).
-                CompressionLevel compressionLevel = this.DefaultCompressionLevel;
-
-                if (mediaRow.HasExplicitCompressionLevel)
-                {
-                    compressionLevel = mediaRow.CompressionLevel;
-                }
-
-                cabinetWorkItem = new CabinetWorkItem(fileRows, resolvedCabinet.Path, maxThreshold, compressionLevel/*, this.FileManager*/);
-            }
-            else // reuse the cabinet from the cabinet cache.
-            {
-                this.core.OnMessage(WixVerboses.ReusingCabCache(mediaRow.SourceLineNumbers, mediaRow.Cabinet, resolvedCabinet.Path));
-
-                try
-                {
-                    // Ensure the cached cabinet timestamp is current to prevent perpetual incremental builds. The
-                    // problematic scenario goes like this. Imagine two cabinets in the cache. Update a file that
-                    // goes into one of the cabinets. One cabinet will get rebuilt, the other will be copied from
-                    // the cache. Now the file (an input) has a newer timestamp than the reused cabient (an output)
-                    // causing the project to look like it perpetually needs a rebuild until all of the reused
-                    // cabinets get newer timestamps.
-                    File.SetLastWriteTime(resolvedCabinet.Path, DateTime.Now);
-                }
-                catch (Exception e)
-                {
-                    this.core.OnMessage(WixWarnings.CannotUpdateCabCache(mediaRow.SourceLineNumbers, resolvedCabinet.Path, e.Message));
-                }
-            }
-
-            if (mediaRow.Cabinet.StartsWith("#", StringComparison.Ordinal))
-            {
-                Table streamsTable = output.EnsureTable(this.core.TableDefinitions["_Streams"]);
-
-                Row streamRow = streamsTable.CreateRow(null);
-                streamRow[0] = mediaRow.Cabinet.Substring(1);
-                streamRow[1] = resolvedCabinet.Path;
-            }
-            else
-            {
-                string destinationPath = Path.Combine(cabinetDir, mediaRow.Cabinet);
-                FileTransfer transfer;
-                if (FileTransfer.TryCreate(resolvedCabinet.Path, destinationPath, CabinetBuildOption.BuildAndMove == resolvedCabinet.BuildOption, "Cabinet", mediaRow.SourceLineNumbers, out transfer))
-                {
-                    transfer.Built = true;
-                    fileTransfers.Add(transfer);
-                }
-            }
-
-            return cabinetWorkItem;
         }
 
         /// <summary>
@@ -7042,46 +6886,6 @@ namespace WixToolset
                             }
                         }
                     }
-                }
-            }
-        }
-
-        /// <summary>
-        /// Sets the thead count to the number of processors if the current thread count is set to 0.
-        /// </summary>
-        /// <remarks>The thread count value must be greater than 0 otherwise and exception will be thrown.</remarks>
-        private void SetCabbingThreadCount()
-        {
-            // default the number of cabbing threads to the number of processors if it wasn't specified
-            if (0 == this.CabbingThreadCount)
-            {
-                string numberOfProcessors = System.Environment.GetEnvironmentVariable("NUMBER_OF_PROCESSORS");
-
-                try
-                {
-                    if (null != numberOfProcessors)
-                    {
-                        this.CabbingThreadCount = Convert.ToInt32(numberOfProcessors, CultureInfo.InvariantCulture.NumberFormat);
-
-                        if (0 >= this.CabbingThreadCount)
-                        {
-                            throw new WixException(WixErrors.IllegalEnvironmentVariable("NUMBER_OF_PROCESSORS", numberOfProcessors));
-                        }
-                    }
-                    else // default to 1 if the environment variable is not set
-                    {
-                        this.CabbingThreadCount = 1;
-                    }
-
-                    this.core.OnMessage(WixVerboses.SetCabbingThreadCount(this.CabbingThreadCount.ToString()));
-                }
-                catch (ArgumentException)
-                {
-                    throw new WixException(WixErrors.IllegalEnvironmentVariable("NUMBER_OF_PROCESSORS", numberOfProcessors));
-                }
-                catch (FormatException)
-                {
-                    throw new WixException(WixErrors.IllegalEnvironmentVariable("NUMBER_OF_PROCESSORS", numberOfProcessors));
                 }
             }
         }
@@ -7310,80 +7114,6 @@ namespace WixToolset
             }
         }
 
-        /// <summary>
-        /// Gets Compiler Values of MediaTemplate Attributes governing Maximum Cabinet Size after applying Environment Variable Overrides
-        /// </summary>
-        /// <param name="output">Output to generate image for.</param>
-        /// <param name="fileRows">The indexed file rows.</param>
-        private void GetMediaTemplateAttributes(out int maxCabSizeForLargeFileSplitting, out int maxUncompressedMediaSize)
-        {
-            // Get Environment Variable Overrides for MediaTemplate Attributes governing Maximum Cabinet Size
-            string mcslfsString = Environment.GetEnvironmentVariable("WIX_MCSLFS");
-            string mumsString = Environment.GetEnvironmentVariable("WIX_MUMS");
-            int maxCabSizeForLargeFileInMB = 0;
-            int maxPreCompressedSizeInMB = 0;
-            ulong testOverFlow = 0;
-
-            // Supply Compile MediaTemplate Attributes to Cabinet Builder
-            Table mediaTemplateTable = this.output.Tables["WixMediaTemplate"];
-            if (mediaTemplateTable != null)
-            {
-                WixMediaTemplateRow mediaTemplateRow = (WixMediaTemplateRow)mediaTemplateTable.Rows[0];
-
-                // Get the Value for Max Cab Size for File Splitting
-                try
-                {
-                    // Override authored mcslfs value if environment variable is authored.
-                    if (!String.IsNullOrEmpty(mcslfsString))
-                    {
-                        maxCabSizeForLargeFileInMB = Int32.Parse(mcslfsString);
-                    }
-                    else
-                    {
-                        maxCabSizeForLargeFileInMB = mediaTemplateRow.MaximumCabinetSizeForLargeFileSplitting;
-                    }
-                    testOverFlow = (ulong)maxCabSizeForLargeFileInMB * 1024 * 1024;
-                }
-                catch (FormatException)
-                {
-                    throw new WixException(WixErrors.IllegalEnvironmentVariable("WIX_MCSLFS", mcslfsString));
-                }
-                catch (OverflowException)
-                {
-                    throw new WixException(WixErrors.MaximumCabinetSizeForLargeFileSplittingTooLarge(null, maxCabSizeForLargeFileInMB, CompilerCore.MaxValueOfMaxCabSizeForLargeFileSplitting));
-                }
-
-                try
-                {
-                    // Override authored mums value if environment variable is authored.
-                    if (!String.IsNullOrEmpty(mumsString))
-                    {
-                        maxPreCompressedSizeInMB = Int32.Parse(mumsString);
-                    }
-                    else
-                    {
-                        maxPreCompressedSizeInMB = mediaTemplateRow.MaximumUncompressedMediaSize;
-                    }
-                    testOverFlow = (ulong)maxPreCompressedSizeInMB * 1024 * 1024;
-                }
-                catch (FormatException)
-                {
-                    throw new WixException(WixErrors.IllegalEnvironmentVariable("WIX_MUMS", mumsString));
-                }
-                catch (OverflowException)
-                {
-                    throw new WixException(WixErrors.MaximumUncompressedMediaSizeTooLarge(null, maxPreCompressedSizeInMB));
-                }
-
-                maxCabSizeForLargeFileSplitting = maxCabSizeForLargeFileInMB;
-                maxUncompressedMediaSize = maxPreCompressedSizeInMB;
-            }
-            else
-            {
-                maxCabSizeForLargeFileSplitting = 0;
-                maxUncompressedMediaSize = CompilerCore.DefaultMaximumUncompressedMediaSize;
-            }
-        }
 
         /// <summary>
         /// The types that the WixPatchSymbolPaths table can hold (and that the WixPatchSymbolPathsComparer can sort).
@@ -7541,142 +7271,5 @@ namespace WixToolset
             // Defaults to the bundle ID as the provider key.
         }
         #endregion
-
-        /// <summary>
-        /// Delegate for Cabinet Split Callback
-        /// </summary>
-        [UnmanagedFunctionPointer(CallingConvention.StdCall)]
-        internal delegate void FileSplitCabNamesCallback([MarshalAs(UnmanagedType.LPWStr)]string firstCabName, [MarshalAs(UnmanagedType.LPWStr)]string newCabName, [MarshalAs(UnmanagedType.LPWStr)]string fileToken);
-
-        /// <summary>
-        /// Call back to Add File Transfer for new Cab and add new Cab to Media table
-        /// This callback can come from Multiple Cabinet Builder Threads and so should be thread safe
-        /// This callback will not be called in case there is no File splitting. i.e. MaximumCabinetSizeForLargeFileSplitting was not authored
-        /// </summary>
-        /// <param name="firstCabName">The name of splitting cabinet without extention e.g. "cab1".</param>
-        /// <param name="newCabName">The name of the new cabinet that would be formed by splitting e.g. "cab1b.cab"</param>
-        /// <param name="fileToken">The file token of the first file present in the splitting cabinet</param>
-        internal void NewCabNamesCallBack([MarshalAs(UnmanagedType.LPWStr)]string firstCabName, [MarshalAs(UnmanagedType.LPWStr)]string newCabName, [MarshalAs(UnmanagedType.LPWStr)]string fileToken)
-        {
-            // Locking Mutex here as this callback can come from Multiple Cabinet Builder Threads
-            Mutex mutex = new Mutex(false, "WixCabinetSplitBinderCallback");
-            try
-            {
-                if (!mutex.WaitOne(0, false)) // Check if you can get the lock
-                {
-                    // Cound not get the Lock
-                    this.core.OnMessage(WixVerboses.CabinetsSplitInParallel());
-                    mutex.WaitOne(); // Wait on other thread
-                }
-
-                string firstCabinetName = firstCabName + ".cab";
-                string newCabinetName = newCabName;
-                bool transferAdded = false; // Used for Error Handling
-
-                // Create File Transfer for new Cabinet using transfer of Base Cabinet
-                foreach (FileTransfer transfer in this.fileTransfers)
-                {
-                    if (firstCabinetName.Equals(Path.GetFileName(transfer.Source), StringComparison.InvariantCultureIgnoreCase))
-                    {
-                        string newCabSourcePath = Path.Combine(Path.GetDirectoryName(transfer.Source), newCabinetName);
-                        string newCabTargetPath = Path.Combine(Path.GetDirectoryName(transfer.Destination), newCabinetName);
-
-                        FileTransfer newTransfer;
-                        if (FileTransfer.TryCreate(newCabSourcePath, newCabTargetPath, transfer.Move, "Cabinet", transfer.SourceLineNumbers, out newTransfer))
-                        {
-                            newTransfer.Built = true;
-                            this.fileTransfers.Add(newTransfer);
-                            transferAdded = true;
-                            break;
-                        }
-                    }
-                }
-
-                // Check if File Transfer was added
-                if (!transferAdded)
-                {
-                    throw new WixException(WixErrors.SplitCabinetCopyRegistrationFailed(newCabinetName, firstCabinetName));
-                }
-
-                // Add the new Cabinets to media table using LastSequence of Base Cabinet
-                Table mediaTable = this.output.Tables["Media"];
-                Table fileTable = this.output.Tables["File"];
-                int diskIDForLastSplitCabAdded = 0; // The DiskID value for the first cab in this cabinet split chain
-                int lastSequenceForLastSplitCabAdded = 0; // The LastSequence value for the first cab in this cabinet split chain
-                bool lastSplitCabinetFound = false; // Used for Error Handling
-
-                string lastCabinetOfThisSequence = String.Empty;
-                // Get the Value of Last Cabinet Added in this split Sequence from Dictionary
-                if (!this.lastCabinetAddedToMediaTable.TryGetValue(firstCabinetName, out lastCabinetOfThisSequence))
-                {
-                    // If there is no value for this sequence, then use first Cabinet is the last one of this split sequence
-                    lastCabinetOfThisSequence = firstCabinetName;
-                }
-
-                foreach (MediaRow mediaRow in mediaTable.Rows)
-                {
-                    // Get details for the Last Cabinet Added in this Split Sequence
-                    if ((lastSequenceForLastSplitCabAdded == 0) && lastCabinetOfThisSequence.Equals(mediaRow.Cabinet, StringComparison.InvariantCultureIgnoreCase))
-                    {
-                        lastSequenceForLastSplitCabAdded = mediaRow.LastSequence;
-                        diskIDForLastSplitCabAdded = mediaRow.DiskId;
-                        lastSplitCabinetFound = true;
-                    }
-
-                    // Check for Name Collision for the new Cabinet added
-                    if (newCabinetName.Equals(mediaRow.Cabinet, StringComparison.InvariantCultureIgnoreCase))
-                    {
-                        // Name Collision of generated Split Cabinet Name and user Specified Cab name for current row
-                        throw new WixException(WixErrors.SplitCabinetNameCollision(newCabinetName, firstCabinetName));
-                    }
-                }
-
-                // Check if the last Split Cabinet was found in the Media Table
-                if (!lastSplitCabinetFound)
-                {
-                    throw new WixException(WixErrors.SplitCabinetInsertionFailed(newCabinetName, firstCabinetName, lastCabinetOfThisSequence));
-                }
-
-                // The new Row has to be inserted just after the last cab in this cabinet split chain according to DiskID Sort
-                // This is because the FDI Extract requires DiskID of Split Cabinets to be continuous. It Fails otherwise with 
-                // Error 2350 (FDI Server Error) as next DiskID did not have the right split cabinet during extraction
-                MediaRow newMediaRow = (MediaRow)mediaTable.CreateRow(null);
-                newMediaRow.Cabinet = newCabinetName;
-                newMediaRow.DiskId = diskIDForLastSplitCabAdded + 1; // When Sorted with DiskID, this new Cabinet Row is an Insertion
-                newMediaRow.LastSequence = lastSequenceForLastSplitCabAdded;
-
-                // Now increment the DiskID for all rows that come after the newly inserted row to Ensure that DiskId is unique
-                foreach (MediaRow mediaRow in mediaTable.Rows)
-                {
-                    // Check if this row comes after inserted row and it is not the new cabinet inserted row
-                    if (mediaRow.DiskId >= newMediaRow.DiskId && !newCabinetName.Equals(mediaRow.Cabinet, StringComparison.InvariantCultureIgnoreCase))
-                    {
-                        mediaRow.DiskId++; // Increment DiskID
-                    }
-                }
-
-                // Now Increment DiskID for All files Rows so that the refer to the right Media Row
-                foreach (FileRow fileRow in fileTable.Rows)
-                {
-                    // Check if this row comes after inserted row and if this row is not the file that has to go into the current cabinet
-                    // This check will work as we have only one large file in every splitting cabinet
-                    // If we want to support splitting cabinet with more large files we need to update this code
-                    if (fileRow.DiskId >= newMediaRow.DiskId && !fileRow.File.Equals(fileToken, StringComparison.InvariantCultureIgnoreCase))
-                    {
-                        fileRow.DiskId++; // Increment DiskID
-                    }
-                }
-
-                // Update the Last Cabinet Added in the Split Sequence in Dictionary for future callback
-                this.lastCabinetAddedToMediaTable[firstCabinetName] = newCabinetName;
-
-                mediaTable.ValidateRows(); // Valdiates DiskDIs, throws Exception as Wix Error if validation fails
-            }
-            finally
-            {
-                // Releasing the Mutex here
-                mutex.ReleaseMutex();
-            }
-        }
     }
 }
