+++ src/tools/wix/Wix.csproj
@@ -24,13 +24,14 @@
     <Compile Include="BinderCore.cs" />
     <Compile Include="BinderFileManagerCore.cs" />
     <Compile Include="BinderFileManager.cs" />
+    <Compile Include="Bind\CreateCabinetsCommand.cs" />
     <Compile Include="BundleChainAttributes.cs" />
     <Compile Include="BundlePackageAttributes.cs" />
     <Compile Include="BurnCommon.cs" />
     <Compile Include="BurnReader.cs" />
     <Compile Include="BurnWriter.cs" />
-    <Compile Include="CabinetBuilder.cs" />
-    <Compile Include="CabinetWorkItem.cs" />
+    <Compile Include="Bind\CabinetBuilder.cs" />
+    <Compile Include="Bind\CabinetWorkItem.cs" />
     <Compile Include="Cab\CabinetFileInfo.cs" />
     <Compile Include="Cab\Interop\CabInterop.cs" />
     <Compile Include="Cab\WixCreateCab.cs" />
